QR scan implemented
